submit verify feature add

+++ submit/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import user_passes_test
+from .utils import staff_or_valid_session_check, user_passes_test_with_request, staff_check
 from django.core.paginator import Paginator
 from .forms import (JmSubmitForm, PersonForm)
 from .models import Submit, Person
@@ -40,7 +42,41 @@ def jm_submit(request):
 def submit_detail(request,pk):
     submit = get_object_or_404(Submit, pk=pk)
     person = Person.objects.get(submit__id=submit.id)
-    return render(request, 'submit/submit_detail.html', {'submit': submit, 'person': person})
+    if request.user.is_authenticated and request.user.is_staff:
+        return render(request, 'submit/submit_detail.html', {'submit': submit, 'person': person})    
+    
+    name = request.session.get('submit_name')
+    phone = request.session.get('submit_phone')
+    if not name or not phone:
+        return redirect('submit-verify')
+    if not (request.user.is_staff or name == submit.name):
+        return redirect('home')
+    
+    else:
+        return render(request, 'submit/submit_detail.html', {'submit': submit, 'person': person}) 
+
+def submit_verify(request):
+    if request.method == "POST":
+        submit_name = request.POST.get('submit_name')
+        submit_phone = request.POST.get('submit_phone')
+        submits = Submit.objects.filter(name=submit_name, phone=submit_phone)
+        # 객체 수에 따른 조건 분기
+        if not submits:
+            # 쿼리셋이 비어 있을 경우
+            return render(request,'submit/submit_verify.html',{'no_submit': True})  
+        elif len(submits) == 1:
+            # 쿼리셋에 객체가 하나만 있는 경우
+            submit = submits.first()
+            person = Person.objects.get(submit__id=submit.id)
+            return render(request,'submit/submit_detail.html',{'submit': submit, 'person': person}) 
+        else:
+            # 쿼리셋에 여러 객체가 있는 경우
+            request.session['submit_name'] = submit_name
+            request.session['submit_phone'] = submit_phone
+            return render(request, 'submit/submit_verify_list.html',{'submits':submits})
+    else:
+        return render(request, 'submit/submit_verify.html')
+
 
 
 def gm_submit(request):
@@ -52,6 +88,8 @@ def sj_submit(request):
 def etc_submit(request):
     return render(request, 'submit/etc_submit.html')
 
+
+@user_passes_test(staff_check, login_url='/')
 def submit_list(request, status=''):
     if status:
         submits = Submit.objects.filter(process=status)
@@ -75,9 +113,14 @@ def submit_list(request, status=''):
 
     return render(request, 'submit/list.html', context)
 
+@user_passes_test_with_request(staff_or_valid_session_check)
 def submit_edit(request,pk):
     submit = get_object_or_404(Submit, pk=pk)
     person = Person.objects.get(submit__id=submit.id)
+    session_name = request.session.get('submit_name', None)
+    if not (request.user.is_staff or (session_name and session_name == submit.name)):
+        return redirect('home')
+        
     if request.method == 'POST':
         submitForm = JmSubmitForm(request.POST,instance=submit)
         personForm = PersonForm(request.POST,instance=person)
@@ -91,7 +134,11 @@ def submit_edit(request,pk):
         context = {'submit': submitForm, 'person': personForm}
     return render(request, 'submit/jm_submit.html', context)
 
+@user_passes_test_with_request(staff_or_valid_session_check)
 def submit_delete(request,pk):
     submit = get_object_or_404(Submit, pk=pk)
+    session_name = request.session.get('submit_name', None)
+    if not (request.user.is_staff or (session_name and session_name == submit.name)):
+        return redirect('home')
     submit.delete()
     return redirect('submit-list')
